gen: Populate keywords in package.json
Even if no input is provided, the filter returns
an array.

Avoid running through the loop, if it contains an
empty string.

This situation arises only when no input was provided.
Regex substitutions should prevent any recurrence of
empty keyword.

It appears that location of "<%=" and "-%>" have an
effect on placement of strings during substitution.

Skipped code optimization for said conditionds in
favor of readability and final indentation of the
keywords - arrived after number of trials.

Signed-off-by: Sanjeev Premi <spremi@ymail.com>