tpl: grunt: Remove leading './' from base paths used
The base paths defined as self.{src,dst,tmp} were
explicitly set to be relative to the module base.

Definition of 'self.src' is used as-is in the
task 'ngTemplates'.

This causes the tag in $templateCache.put()
to be different from path specified in the
templateUrl within the directive.

The application would then fail to load the
templates contained in the module.

Removed leading './' from all path definitions
for consistency.

Signed-off-by: Sanjeev Premi <spremi@ymail.com>